initial work towards eliminating hard-coded level scenes
Eliminated the unnecessary "square" class, which was just a wrapper for
squaretype. Changed the "grid" from an array to a Dictionary wrapper
that never gives out-of-bounds exceptions, so that levels can be
arbitrarily sized and shaped. Level size/shape should be dictated by a
boundary of obstacles rather than by built-in limitation, I think. The
level editor can impose the condition that the level needs to be "closed
up," when we reach that point.)

Created the LevelSkeleton class which records all necessary information
about a level, and which can load that information from an XML file.

Created the LevelScript class which generates a LevelSkeleton from a
given XML file and then loads all the necessary components into the
scene.

Created an XMLLevel scene which allows us to test this with arbitrary
.xml files.

Started changing some of the appearance, don't love the changes, broke
some animations.

TODO:
- Add exotic bricks to XML
- Add next-level logic to the XML thing
- Convert levels to XML
- Redesign all bricks and animations
- Level editor!